+++ src/main/java/screen/AT2ACCSU0018/AT2ACCSU0018Test.java
@@ -10,7 +10,7 @@ import org.apache.bcel.generic.IF_ACMPEQ;
 
 /**
  * Created by vsolis on 29/09/2016.
- * <p>
+ *
  * Bug en el campo hotel description, se borra en la tabla nada mas darle al search
  */
 public class AT2ACCSU0018Test {
@@ -183,7 +183,7 @@ public class AT2ACCSU0018Test {
             return false;
         }
         if (!Functions.insertInput(driver, new String[]{"hotel_comments_qbe_hotel_description", getElements("hotel_comments_qbe_hotel_description")},
-                "qbe_hotel_description", data.getData().get("hotel_description"), " on QBE")) {
+                "qbe_hotel_description", "%" + data.getData().get("hotel_description"), " on QBE")) {
             return false;
         }
         if (!Functions.insertInput(driver, new String[]{"hotel_comments_qbe_chain", getElements("hotel_comments_qbe_chain")},
@@ -487,7 +487,11 @@ public class AT2ACCSU0018Test {
     private boolean Delete_multi(TestDriver driver) {
         driver.getReport().addHeader("DELETE IN HOTEL COMMENTS", 3, false);
 
-
+        if (!Functions.simpleClick(driver,
+                new String[]{"multi_lenguaje_se_e_result", getElements("multi_lenguaje_se_e_result")}, //element to click
+                " on DELETE")) {
+            return false;
+        }
         if (!Functions.doDeleteNCheck(driver,
                 new String[]{"multi_lenguaje_del_b_delete", getElements("multi_lenguaje_del_b_delete")},
                 new String[]{"multi_lenguaje_del_e_records", getElements("multi_lenguaje_del_e_records")},
@@ -523,6 +527,11 @@ public class AT2ACCSU0018Test {
     private boolean Edit_multi(TestDriver driver) {
         driver.getReport().addHeader(" EDIT IN HOTEL COMMENTS", 3, false);
 
+        if (!Functions.simpleClick(driver,
+                new String[]{"multi_lenguaje_se_e_result", getElements("multi_lenguaje_se_e_result")}, //element to click
+                " on EDIT")) {
+            return false;
+        }
         Functions.break_time(driver, 6, 500);
         if (!Functions.checkClick(driver,
                 new String[]{"multi_lenguaje_ed_b_edit", getElements("multi_lenguaje_ed_b_edit")}, //element to click
@@ -551,6 +560,7 @@ public class AT2ACCSU0018Test {
 
     private boolean Search_multi(TestDriver driver) {
 
+        driver.getReport().addHeader(" SEARCH IN HOTEL COMMENTS", 3, false);
 
         if (!Functions.clickQbE(driver,
                 new String[]{"multi_lenguaje_se_b_qbe", getElements("multi_lenguaje_se_b_qbe")},// query button
@@ -566,7 +576,12 @@ public class AT2ACCSU0018Test {
                 "se_castellano", data.getData().get("castellano"), " on SEARCH")) {
             return false;
         }
-
+        if (!Functions.enterQueryAndClickResult(driver,
+                new String[]{"multi_lenguaje_se_i_remarks", getElements("multi_lenguaje_se_i_remarks")}, //search button
+                new String[]{"multi_lenguaje_se_e_result", getElements("multi_lenguaje_se_e_result")}, //result element
+                " on SEARCH")) {
+            return false;
+        }
         return true;
     }
